apps/desktop/src/python/requirements.txt
# BlackIA Python Dependencies for RAG System
# Requires Python 3.11+

# ===== Text RAG (Sentence Transformers for embeddings) =====
sentence-transformers>=2.2.2

# ===== Colette / ColPali for Vision RAG (Multi-platform: CUDA/MPS/CPU) =====
colpali-engine>=0.3.12
torch>=2.7.0
torchvision>=0.22.0

# ===== MLX Framework (Apple Silicon only - installed automatically by setup script) =====
# mlx==0.20.0
# mlx-lm==0.20.1
# mlx-vlm==0.0.13
# Note: These are automatically installed on arm64 systems by setup-python-venv.sh

# ===== Image Processing =====
Pillow>=11.3.0
pdf2image>=1.17.0

# ===== Vector Store =====
lancedb>=0.15.0
pyarrow>=17.0.0

# ===== Utilities =====
numpy>=1.26.4
python-dotenv>=1.0.1

# ===== Optional: Text Extraction (déjà géré par Node mais utile pour tests) =====
# PyMuPDF==1.24.0  # Alternative plus rapide que pdf2image pour texte seul

# ===== Development (optionnel) =====
# pytest==8.3.0
# black==24.8.0
# mypy==1.11.0

# NOTE: MLX-VLM est automatiquement installé sur Apple Silicon (arm64) via setup-python-venv.sh
# Pour les autres plateformes, Colette/ColPali est la solution recommandée
#
# Installation automatique sur Apple Silicon:
#   npm run setup:python:venv  (dans apps/desktop/)
#   ou
#   bash scripts/setup-python-venv.sh
